Remove PhysicalObject dependency of GameSet
